Poker

Base.:+

Base.:-

Base.:<

Base.:>

Base.isless

Base.isless

Base.isless

Poker.bet

Poker.calculate_odds

Poker.cards

Poker.deal

Poker.get_preflop_odds

Poker.get_return

Poker.get_return

Poker.get_returns

Poker.get_returns

Poker.get_returns_table

Poker.to_string

Base.:+ — Method

+(x, y)

Adds two hands together, combining their cards into a single hand. Probably more akin to a union, but duplicate cards are not supported, which helps catch bugs and ensures hands are exactly the length of the two inputs hands added.

Base.:- — Method

-(x, y)

Subtracts the cards from one hand from another. Subtraction more closely follows set theory's subtraction, as cards that are attempted to be subtracted that aren't in the original hand are ignored. This should probably at least warn the user for most poker-related use cases.

Base.:< — Method

<(x, y)

Syntactic sugar for isless(x, y)

Base.:> — Method

<(x, y)

Syntactic sugar for isless(y, x)

Base.isless — Method

isless(x, y)

Compares if a card is less than another card based on its position in a fresh deck. This isn't actually used in Poker.jl right now, but is probably a handy functionality for any extension of Cards.jl

Base.isless — Method

isless(x, y)

Compares two hands to one another, using Texas Hold'em rules. Ties evaluate as false. If larger hands are provided than five cards, the maximal combination of five cards is used for the comparison.

Base.isless — Method

isless(x, y)

Compares if the suit of a card is ranked lower than that of another card. This isn't actually used in Poker.jl right now, but is probably a handy functionality for any extension of Cards.jl

Poker.bet — Method

bet(dealer, self_strat, opp_strat, self_prob, opp_prob, self_return,
    opp_return, pot, big_blind)

Processes the bets for the players, given thresholds at which each player will not put more money into the game, and the calculated odds of each player winning the showdown. Also returns whether the whole get_return function should return

Poker.calculate_odds — Method

+(self, table[, N])

Computes with N (default 1,000) Monte Carlo simulations the odds of the hand winning the showdown given the cards shown on the table

Poker.cards — Method

cards(hand)

Converts a Hand (from Cards.jl) to an array of cards

Poker.deal — Method

deal(deck, n)

Returns both a new hand of n cards, drawn without replacement from the deck, and the deck without the cards that were drawn

Poker.get_preflop_odds — Method

get_preflop_odds([N])

Computes with N (default 1,000) Monte Carlo simulations the odds of the each starting hand winning the showdown given no cards shown on the table yet. This table is optionally used to speed up game calculations.

Poker.get_return — Method

get_return(self_strat, opp_strat[, preflop_odds_table])

Returns the self return after 1 simulated game, given the multiple thresholds (one for each round of betting; ie, 4) at which either player will not put more money into the game. Optional optimization is available by specifying a pre-calculated preflop odds table

Poker.get_return — Method

get_return(self_strat, opp_strat[, preflop_odds_table])

Returns the self return after 1 simulated game, given the singular thresholds at which either player will not put more money into the game. Optional optimization is available by specifying a pre-calculated preflop odds table

Poker.get_returns — Method

get_returns(self_strat, opp_strat[, preflop_odds_table, N])

Wrapping of get_return, but the mean of N (default 5,000) times

Poker.get_returns — Method

get_returns(self_strat, opp_strat[, preflop_odds_table, N])

Wrapping of get_return, but the mean of N (default 100) times

Poker.get_returns_table — Method

get_returns_table([preflop_odds_table, N])

Calculation of a table of average returns across a gamut of singular threshold strategies, with the optional optimization of pre-calculating the preflop odds table. Default N is 5,000 simulations per pair of strategies (total 50,000,000 games)

Poker.to_string — Method

to_string(hand)

Compares if the suit of a card is ranked lower than that of another card. This isn't actually used in Poker.jl right now, but is probably a handy functionality for any extension of Cards.jl